+++ presentation/src/main/java/team/jsv/icec/ui/main/MainActivity.kt
@@ -6,7 +6,6 @@ import dagger.hilt.android.AndroidEntryPoint
 import team.jsv.icec.base.BaseActivity
 import team.jsv.icec.base.EventObserver
 import team.jsv.icec.base.showToast
-import team.jsv.icec.ui.main.mosaic.MosaicEvent
 import team.jsv.icec.ui.main.mosaic.MosaicViewModel
 import team.jsv.icec.ui.main.mosaic.PictureState
 import team.jsv.icec.ui.main.mosaic.ScreenStep
