+++ presentation/src/main/java/team/jsv/icec/ui/main/mosaic/DetectFaceActivity.kt
@@ -1,14 +1,11 @@
 package team.jsv.icec.ui.main.mosaic
 
-import android.content.Intent
-import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import team.jsv.icec.base.BaseActivity
-import team.jsv.icec.ui.main.MainActivity
 import team.jsv.presentation.R
 import team.jsv.presentation.databinding.ActivityMainMosaicStep1Binding
 
-class MainMosaicStep1 : BaseActivity<ActivityMainMosaicStep1Binding>(R.layout.activity_main_mosaic_step1) {
+class DetectFaceActivity : BaseActivity<ActivityMainMosaicStep1Binding>(R.layout.activity_detect_face) {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
