+++ presentation/src/main/java/team/jsv/icec/ui/main/mosaic/MosaicEvent.kt
