+++ presentation/src/main/res/layout/activity_detect_face.xml
