+++ presentation/src/main/res/navigation/nav_graph.xml
@@ -6,7 +6,7 @@
 
     <fragment
         android:id="@+id/selectMosaicEditFragment"
-        android:name="team.jsv.icec.ui.main.SelectMosaicEditFragment"
+        android:name="team.jsv.icec.ui.main.mosaic.SelectMosaicEditFragment"
         android:label="SelectMosaicEditFragment" >
         <action
             android:id="@+id/action_selectMosaicEditFragment_to_faceSelectFragment"
